ChatbotSuporteP&S/main/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.8
langchain-openai==0.0.5
chromadb==0.4.17
sentence-transformers==2.2.2
pydantic==2.5.0
pydantic-settings==2.1.0
pypdf==3.17.1
python-docx==0.8.11
